♻️ refactor:댓글 메세지 컴포넌트 폴더 이동
